Use code with football-challenge
In a first step, I use my code from task1 with the file from the football-challenge in order to see whats different and at which parts I have to gerneralize.
I had to cheet and look up the columns of "Goals" and "Goals allowed". This will be generalised in a later step.
The difference between the input values (val1 and val2) is now calculated as absolute value.