Defined methods read_data and get_result
I defined the two methods read_data and get_result. I decided to split the functions up (compared to my matlab code) in order to be able to replace components and assure reusability. I might like to use a different method to calculate a result for my data later on.